refactor!: move credential related functionality in one folder
