+++ packages/core/src/credentials/V1/common.ts
