+++ packages/core/src/credentials/attestation/Attestation.ts
@@ -13,7 +13,7 @@ import type {
 } from '@kiltprotocol/types'
 import { DataUtils, SDKErrors } from '@kiltprotocol/utils'
 import * as Did from '@kiltprotocol/did'
-import { DelegationNode } from '../delegation/DelegationNode.js'
+import { DelegationNode } from '../../delegation/DelegationNode.js'
 
 /**
  * An [[Attestation]] certifies a [[Claim]], sent by a claimer in the form of a [[Credential]]. [[Attestation]]s are **written on the blockchain** and are **revocable**.
